// src/types/Employees.ts
export interface IEmployee {
    id: string;
    name: string;
    email: string;
    phone_number: string;
    avatar: string;
    date_of_birth: string;
    marital_status: string;
    gender: string;
    nationality: string;
    address: string;
    employee_type: string;
    working_date: string;
    joining_date: string;
    office_location: string;
    position: string;
    user_id: string;
    department_id: string;
    is_active: boolean;
}